create separate structs for separate views
