+++ internal/views/entrycreate.go
@@ -2,12 +2,15 @@ package views
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"log"
 	"net/http"
 	"net/url"
+	"strings"
 	"time"
 
+	"github.com/Ajnasz/sekret.link/internal/parsers"
 	"github.com/Ajnasz/sekret.link/internal/services"
 	"github.com/Ajnasz/sekret.link/uuid"
 )
@@ -32,38 +35,15 @@ func buildCreatedResponse(meta *services.EntryMeta, keyString string) EntryCreat
 	}
 }
 
-type SecretResponse struct {
-	UUID      string
-	Key       string
-	Data      string
-	Created   time.Time
-	Accessed  time.Time
-	Expire    time.Time
-	DeleteKey string
+type EntryCreateView struct {
+	webExternalURL *url.URL
 }
 
-func buildSecretResponse(meta services.Entry) SecretResponse {
-	return SecretResponse{
-		UUID:      meta.UUID,
-		Created:   meta.Created,
-		Expire:    meta.Expire,
-		Accessed:  meta.Accessed,
-		DeleteKey: meta.DeleteKey,
-		Data:      string(meta.Data),
-	}
+func NewEntryCreateView(webExternalURL *url.URL) EntryCreateView {
+	return EntryCreateView{webExternalURL: webExternalURL}
 }
 
-type EntryView struct {
-	WebExternalURL *url.URL
-}
-
-func NewEntryView(webExternalURL *url.URL) EntryView {
-	return EntryView{
-		WebExternalURL: webExternalURL,
-	}
-}
-
-func (e EntryView) RenderEntryCreated(w http.ResponseWriter, r *http.Request, entry *services.EntryMeta, keyString string) {
+func (e EntryCreateView) RenderEntryCreated(w http.ResponseWriter, r *http.Request, entry *services.EntryMeta, keyString string) {
 	w.Header().Add("x-entry-uuid", entry.UUID)
 	w.Header().Add("x-entry-key", keyString)
 	w.Header().Add("x-entry-expire", entry.Expire.Format(time.RFC3339))
@@ -76,7 +56,7 @@ func (e EntryView) RenderEntryCreated(w http.ResponseWriter, r *http.Request, en
 
 		json.NewEncoder(w).Encode(response)
 	} else {
-		newURL, err := uuid.GetUUIDUrlWithSecret(e.WebExternalURL, entry.UUID, keyString)
+		newURL, err := uuid.GetUUIDUrlWithSecret(e.webExternalURL, entry.UUID, keyString)
 		if err != nil {
 			log.Println("Get UUID URL with secret failed", err)
 			http.Error(w, "Internal error", http.StatusInternalServerError)
@@ -87,18 +67,18 @@ func (e EntryView) RenderEntryCreated(w http.ResponseWriter, r *http.Request, en
 	}
 }
 
-func (e EntryView) RenderReadEntry(w http.ResponseWriter, r *http.Request, entry *services.Entry, keyString string) {
-	if r.Header.Get("Accept") == "application/json" {
-		response := buildSecretResponse(*entry)
-
-		response.Key = keyString
-		json.NewEncoder(w).Encode(response)
+func (e EntryCreateView) RenderCreateEntryErrorResponse(w http.ResponseWriter, r *http.Request, err error) {
+	if errors.Is(err, parsers.ErrInvalidExpirationDate) {
+		http.Error(w, "Invalid expiration", http.StatusBadRequest)
+		return
+	} else if errors.Is(err, parsers.ErrInvalidMaxRead) {
+		http.Error(w, "Invalid max read", http.StatusBadRequest)
+		return
+	} else if errors.Is(err, parsers.ErrInvalidData) {
+		http.Error(w, "Invalid data", http.StatusBadRequest)
+	} else if strings.Contains(err.Error(), "http: request body too large") {
+		http.Error(w, "Too large", http.StatusRequestEntityTooLarge)
 	} else {
-		w.WriteHeader(http.StatusOK)
-		w.Write(entry.Data)
+		http.Error(w, "Internal error", http.StatusInternalServerError)
 	}
-}
-
-func (e EntryView) RenderDeleteEntry(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusAccepted)
 }
